Update README.md
The project focuses on building a mental health tracker. You will try to get an idea of the mental
state of your user (in the least intrusive ways), find out if they are suffering and then suggest
measures they can take to get out of their present condition.
A user answers some questions and based on the answers that they provide, you will suggest tasks
to them and maintain a record of their mental state for displaying on a dashboard. Mental disorders
are widespread in countries all over the world. Nevertheless, there is a global shortage in human
resources delivering mental health services.
Leaving people with mental disorders untreated may increase suicide attempts and mortality. To
address this matter of limited resources, conversational agents have gained momentum in the last
years. In this work, we introduce a mobile application with integrated Chabot that implements
methods from cognitive behavior therapy (CBT) to support mentally ill people in regulating
emotions and dealing with thoughts and feelings.
Application asks the user on daily basis on events that occurred and on emotions. It determines
automatically the basic emotion of a user from the natural language input using natural language
processing and a lexicon-based approach. Depending on the emotion, an appropriate measurement
such as activities or mindfulness exercises is suggested by application.